chore: Move the Event and related types into a library
