+++ lib/event/src/util/log/contains.rs
