+++ lib/event/src/util/log/insert.rs
