+++ lib/event/src/util/log/mod.rs
@@ -8,7 +8,7 @@ mod path_iter;
 mod remove;
 
 pub(self) use super::Value;
-pub(crate) use path_iter::{PathComponent, PathIter};
+pub use path_iter::{PathComponent, PathIter};
 
 pub use all_fields::all_fields;
 pub use contains::contains;
