+++ lib/event/src/util/log/path_iter.rs
