+++ lib/event/src/util/log/remove.rs
